requirements.txt
flask
requests
pytesseract
pillow
openai
opencv-python
numpy
opencv-contrib-python
python-dotenv

